+++ src/compliance_tests/TestIso_8_2_4.cpp
@@ -109,9 +109,9 @@ class TestIso_8_2_4 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 4; i++)
+            for (size_t i = 0; i < 4; i++)
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
-            for (int i = 0; i < 10; i++)
+            for (size_t i = 0; i < 10; i++)
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
 
             SetupMonitorTxTests();
@@ -137,7 +137,7 @@ class TestIso_8_2_4 : public test::TestBase
                     dlc = 0x8;
                     frm_flags = std::make_unique<FrameFlags>(FrameKind::Can20,
                                             IdentKind::Ext, RtrFlag::Data);
-                    for (int i = 0; i < 8; i++)
+                    for (size_t i = 0; i < 8; i++)
                         data[i] = 0x3C;
                     break;
 
@@ -266,7 +266,7 @@ class TestIso_8_2_4 : public test::TestBase
                     break;
                 }
 
-                for (int i = 1; i < 64; i++)
+                for (size_t i = 1; i < 64; i++)
                     data[i] = 0x55;
             }
 
@@ -296,7 +296,7 @@ class TestIso_8_2_4 : public test::TestBase
             drv_bit_frm->UpdateFrame();
             mon_bit_frm->UpdateFrame();
 
-            int num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
+            size_t num_stuff_bits = drv_bit_frm->GetNumStuffBits(StuffKind::Normal);
 
             /* In FD enabled variant, if last bit of data field is stuff bit, but model has this bit
              * as fixed stuff bit before Stuff count. So count in also each fixed stuff bit even
@@ -306,7 +306,7 @@ class TestIso_8_2_4 : public test::TestBase
             if (test_variant == TestVariant::CanFdEna)
             {
                 Bit *bit = drv_bit_frm->GetBitOf(0, BitKind::StuffCnt);
-                int index = drv_bit_frm->GetBitIndex(bit);
+                size_t index = drv_bit_frm->GetBitIndex(bit);
                 BitVal value = drv_bit_frm->GetBit(index - 1)->val_;
                 if ((value == drv_bit_frm->GetBit(index - 2)->val_) &&
                     (value == drv_bit_frm->GetBit(index - 3)->val_) &&
@@ -318,9 +318,9 @@ class TestIso_8_2_4 : public test::TestBase
             /**************************************************************************************
              * Execute test
              *************************************************************************************/
-            for (int stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
+            for (size_t stuff_bit = 0; stuff_bit < num_stuff_bits; stuff_bit++)
             {
-                TestMessage("Testing stuff bit nr: %d", stuff_bit);
+                TestMessage("Testing stuff bit nr: %zu", stuff_bit);
                 stuff_bits_in_variant++;
 
                 /*
@@ -331,7 +331,7 @@ class TestIso_8_2_4 : public test::TestBase
                 mon_bit_frm_2 = std::make_unique<BitFrame>(*mon_bit_frm);
 
                 Bit *stuff_bit_to_flip = drv_bit_frm_2->GetStuffBit(stuff_bit);
-                int bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
+                size_t bit_index = drv_bit_frm_2->GetBitIndex(stuff_bit_to_flip);
                 stuff_bit_to_flip->FlipVal();
 
                 if (is_err_passive)
